requirements.txt
pandas
mysql-connector-python
chardet